requirements.txt
deeplake
langchain
openai
pathspec
python-dotenv
streamlit
streamlit_chat
tiktoken
